refactor(core): extract FileWindows from the window geometry
Move the container-backed streaming decoder into stream/file.rs.
WindowSpec, SliceWindows, and the PcmWindows trait stay in stream.rs.
